% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_auth.R
\name{apl_get_access_token}
\alias{apl_get_access_token}
\title{Get access_token}
\usage{
apl_get_access_token(
  client_id,
  jwt_data,
  account_name,
  cache_path = rappdirs::site_data_dir("rappleads")
)
}
\arguments{
\item{client_id}{You receive your clientId when you upload a public key.}

\item{jwt_data}{JWT object \code{apl_get_client_secret()}}

\item{account_name}{Your apple ads account name}

\item{cache_path}{Path to the directory where cached authentication data will be stored.}
}
\value{
access_token object
}
\description{
Get access_token
}
