% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_get_ad_groups.R
\name{apl_get_ad_groups}
\alias{apl_get_ad_groups}
\title{Get Ad Groups}
\usage{
apl_get_ad_groups(org_id = apl_get_me_details()$parentOrgId)
}
\arguments{
\item{org_id}{The value is your orgId.}
}
\value{
tibble with ad group metadata
}
\description{
Get Ad Groups
}
