% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_get_ads.R
\name{apl_get_ads}
\alias{apl_get_ads}
\title{Get Ads}
\usage{
apl_get_ads(org_id = apl_get_me_details()$parentOrgId)
}
\arguments{
\item{org_id}{The value is your orgId.}
}
\value{
tibble with ads metadata
}
\description{
Get Ads
}
