% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{apl_get_auth_account_list}
\alias{apl_get_auth_account_list}
\title{Get List of Auth Cached Accounts}
\usage{
apl_get_auth_account_list()
}
\value{
character vector with accounts name
}
\description{
Get List of Auth Cached Accounts
}
