% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_get_budget_orders.R
\name{apl_get_budget_orders}
\alias{apl_get_budget_orders}
\title{Get All Budget Orders}
\usage{
apl_get_budget_orders(org_id)
}
\arguments{
\item{org_id}{The value is your orgId.}
}
\value{
tibble with budget data
}
\description{
Get All Budget Orders
}
