% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_get_campaigns.R
\name{apl_get_campaigns}
\alias{apl_get_campaigns}
\title{Get all Campaigns}
\usage{
apl_get_campaigns(org_id)
}
\arguments{
\item{org_id}{The value is your orgId.}
}
\value{
tibble with campaigns
}
\description{
Get all Campaigns
}
