% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_auth.R
\name{apl_get_client_secret}
\alias{apl_get_client_secret}
\title{Get client secret}
\usage{
apl_get_client_secret(
  client_id,
  team_id,
  key_id,
  private_key_path,
  account_name,
  cache_path = rappdirs::site_data_dir("rappleads")
)
}
\arguments{
\item{client_id}{You receive your clientId when you upload a public key.}

\item{team_id}{The client secret is a JWT that you create and sign with your private key.}

\item{key_id}{The value is your keyId that returns when you upload a public key.}

\item{private_key_path}{Path to the \code{.pem} file containing your private key.}

\item{account_name}{Your apple ads account name}

\item{cache_path}{Path to the directory where cached authentication data will be stored.}
}
\value{
jwt_data
}
\description{
Get client secret
}
