% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_get_creatives.R
\name{apl_get_creatives}
\alias{apl_get_creatives}
\title{Get All Creatives}
\usage{
apl_get_creatives(org_id = apl_get_me_details()$parentOrgId)
}
\arguments{
\item{org_id}{The value is your orgId.}
}
\value{
tibble with campaigns
}
\description{
Get All Creatives
}
