% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apl_make_request}
\alias{apl_make_request}
\title{Make API request}
\usage{
apl_make_request(endpoint, parser, org_id = NULL, selector = NULL)
}
\arguments{
\item{endpoint}{API endpoint}

\item{parser}{Parser function}

\item{org_id}{Organization ID}

\item{selector}{\href{https://developer.apple.com/documentation/apple_ads/selector}{Selector} objects define what data the API returns when fetching resources.}
}
\value{
API parsed response
}
\description{
Make API request
}
