% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{apl_set_account_name}
\alias{apl_set_account_name}
\title{Set Apple Ads Account Name}
\usage{
apl_set_account_name(account_name)
}
\arguments{
\item{account_name}{Your apple ads account name}
}
\value{
nothing
}
\description{
Set Apple Ads Account Name
}
