% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{apl_set_client_id}
\alias{apl_set_client_id}
\title{Set Client ID}
\usage{
apl_set_client_id(client_id)
}
\arguments{
\item{client_id}{You receive your clientId when you upload a public key.}
}
\value{
nothing
}
\description{
Set Client ID
}
