% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{apl_set_key_id}
\alias{apl_set_key_id}
\title{Set KeyID}
\usage{
apl_set_key_id(key_id)
}
\arguments{
\item{key_id}{The value is your keyId that returns when you upload a public key.}
}
\value{
nothing
}
\description{
Set KeyID
}
