% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{apl_set_private_key_path}
\alias{apl_set_private_key_path}
\title{Set Privat Key Path}
\usage{
apl_set_private_key_path(private_key_path)
}
\arguments{
\item{private_key_path}{Path to the \code{.pem} file containing your private key.}
}
\value{
nothing
}
\description{
Set Privat Key Path
}
