% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{apl_set_team_id}
\alias{apl_set_team_id}
\title{Set Team ID}
\usage{
apl_set_team_id(team_id)
}
\arguments{
\item{team_id}{The client secret is a JWT that you create and sign with your private key.}
}
\value{
nothing
}
\description{
Set Team ID
}
