% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{apl_set_timezone}
\alias{apl_set_timezone}
\title{Set Tinezone}
\usage{
apl_set_timezone(timezone)
}
\arguments{
\item{timezone}{You set the default timeZone during account creation through the Apple Search Ads UI. ORTZ (organization time zone) is the default. Possible Values: ORTZ, UTC}
}
\value{
nothing
}
\description{
Set Tinezone
}
