% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_selector}
\alias{make_selector}
\title{Make request selector}
\usage{
make_selector(
  start_date = Sys.Date() - 8,
  end_date = Sys.Date() - 1,
  granularity = c("DAILY", "HOURLY", "WEEKLY", "MONTHLY"),
  sort_field = "startTime",
  time_zone = "UTC",
  group_by = NULL,
  part = NULL
)
}
\arguments{
\item{start_date}{start report date}

\item{end_date}{end report date}

\item{granularity}{The report data organized by hour, day, week, and month.}

\item{sort_field}{Specify a field name and grouping to sort the records by ASCENDING or DESCENDING.}

\item{time_zone}{You set the default timeZone during account creation through the Apple Search Ads UI. ORTZ}

\item{group_by}{Use the groupBy field to group responses by selected dimensions. If groupBy specifies age, gender, and geodimensions}

\item{part}{Part of selector}
}
\description{
Make request selector
}
\details{
See \href{https://developer.apple.com/documentation/apple_ads/reportingrequest}{API Documentation}
}
