% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rappleads-package.R
\docType{package}
\name{rappleads-package}
\alias{rappleads}
\alias{rappleads-package}
\title{rappleads: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selesnow/rappleads}
  \item Report bugs at \url{https://github.com/selesnow/rappleads/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
