% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rappleads-package.R
\docType{package}
\name{rappleads-package}
\alias{rappleads}
\alias{rappleads-package}
\title{rappleads: Get Data From 'Apple Ads Campaign Management API'}
\description{
Load data from Apple Ads Accounts using 'Apple Ads Campaign Management API' \url{https://developer.apple.com/documentation/apple_ads}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selesnow/rappleads}
  \item Report bugs at \url{https://github.com/selesnow/rappleads/issues}
}

}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

Other contributors:
\itemize{
  \item Netpeak [copyright holder]
}

}
\keyword{internal}
